# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RevisionConflictDetails"
              },
              {
                "type": "null"
              }
            ]
          },
          "field_errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Field Errors",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "message_args": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Message Args",
            "type": "object"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "ErrorCode": {
        "enum": [
          "BAD_REQUEST",
          "INPUT_INVALID",
          "AUTH_REQUIRED",
          "AUTH_LOGIN_FAILED",
          "ACCESS_EXPIRED",
          "SESSION_REVOKED",
          "SESSION_INVALID",
          "PERMISSION_DENIED",
          "CSRF_FAILED",
          "RESOURCE_NOT_FOUND",
          "METHOD_NOT_ALLOWED",
          "REVISION_CONFLICT",
          "IDEMPOTENCY_CONFLICT",
          "STATE_CONFLICT",
          "KEY_REQUIRED",
          "REFRESH_SUPERSEDED",
          "RESOURCE_EXPIRED",
          "PAYLOAD_TOO_LARGE",
          "MEDIA_TYPE_UNSUPPORTED",
          "CAPABILITY_UNSUPPORTED",
          "RATE_LIMITED",
          "QUOTA_EXCEEDED",
          "INTERNAL_ERROR",
          "DEPENDENCY_ERROR",
          "SERVICE_UNAVAILABLE",
          "DEPENDENCY_TIMEOUT",
          "EXTERNAL_RESULT_UNKNOWN"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "error",
          "meta"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "FieldError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/FieldErrorCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "message_args": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Message Args",
            "type": "object"
          },
          "path": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Path",
            "type": "array"
          },
          "source": {
            "enum": [
              "body",
              "query",
              "path",
              "header",
              "cookie"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "source",
          "path",
          "code",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
      "FieldErrorCode": {
        "enum": [
          "VALIDATION_REQUIRED",
          "VALIDATION_TYPE",
          "VALIDATION_FORMAT",
          "VALIDATION_TOO_LONG",
          "VALIDATION_OUT_OF_RANGE",
          "VALIDATION_UNKNOWN_FIELD",
          "VALIDATION_INVALID"
        ],
        "title": "FieldErrorCode",
        "type": "string"
      },
...
      "ResponseMeta": {
        "additionalProperties": false,
        "properties": {
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id"
        ],
        "title": "ResponseMeta",
        "type": "object"
      },
      "RevisionConflictDetails": {
        "additionalProperties": false,
        "properties": {
          "current_revision": {
            "minimum": 0.0,
            "title": "Current Revision",
            "type": "integer"
          },
          "kind": {
            "const": "revision_conflict",
            "default": "revision_conflict",
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "current_revision"
        ],
        "title": "RevisionConflictDetails",
        "type": "object"
      },
...
      "SuccessResponse_TelemetryBatchRead_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TelemetryBatchRead"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[TelemetryBatchRead]",
        "type": "object"
      },
...
      "TelemetryBatchRead": {
        "additionalProperties": false,
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/TelemetryItemResult"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "title": "TelemetryBatchRead",
        "type": "object"
      },
      "TelemetryItemResult": {
        "additionalProperties": false,
        "properties": {
          "event_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "index": {
            "minimum": 0.0,
            "title": "Index",
            "type": "integer"
          },
          "reason": {
            "anyOf": [
              {
                "enum": [
                  "invalid_event",
                  "invalid_attributes",
                  "invalid_time",
                  "invalid_platform",
                  "invalid_record",
                  "not_allowed_for_audience"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "enum": [
              "accepted",
              "duplicate",
              "rejected"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "index",
          "event_id",
          "status"
        ],
        "title": "TelemetryItemResult",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Haruka API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/frontend-logs": {
      "post": {
        "operationId": "receive_admin_telemetry",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "description": "OpenAPI item shape; runtime parses raw items for independent rejects.",
                "properties": {
                  "client_session_id": {
                    "format": "uuid4",
                    "title": "Client Session Id",
                    "type": "string"
                  },
                  "events": {
                    "items": {
                      "additionalProperties": false,
                      "properties": {
                        "attributes": {
                          "anyOf": [
                            {
                              "additionalProperties": false,
                              "description": "Only low-risk categories/counts; no text, email, URL, token or resource ID.",
                              "properties": {
                                "card_type": {
                                  "anyOf": [
                                    {
                                      "const": "word",
                                      "type": "string"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ],
                                  "title": "Card Type"
                                },
                                "drop_reason": {
                                  "anyOf": [
                                    {
                                      "enum": [
                                        "queue_full",
                                        "expired",
                                        "auth_scope_changed",
                                        "upload_failure",
                                        "invalid_event",
                                        "storage_unavailable"
                                      ],
                                      "type": "string"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ],
                                  "title": "Drop Reason"
                                },
                                "dropped_count": {
                                  "anyOf": [
                                    {
                                      "maximum": 5000.0,
                                      "minimum": 0.0,
                                      "type": "integer"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ],
                                  "title": "Dropped Count"
                                },
                                "duration_ms": {
                                  "anyOf": [
                                    {
                                      "maximum": 86400000.0,
                                      "minimum": 0.0,
                                      "type": "number"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ],
                                  "title": "Duration Ms"
                                },
                                "error_category": {
                                  "anyOf": [
                                    {
                                      "enum": [
                                        "authentication",
                                        "authorization",
                                        "conflict",
                                        "network",
                                        "timeout",
                                        "server",
                                        "validation",
                                        "flutter_framework",
                                        "dart_unhandled",
                                        "platform_channel",
                                        "other"
                                      ],
                                      "type": "string"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ],
                                  "title": "Error Category"
                                },
                                "material_type": {
                                  "anyOf": [
                                    {
                                      "const": "novel",
                                      "type": "string"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ],
                                  "title": "Material Type"
                                },
                                "result": {
                                  "anyOf": [
                                    {
                                      "enum": [
                                        "success",
                                        "failure",
                                        "cancelled",
                                        "denied",
                                        "hit",
                                        "miss"
                                      ],
                                      "type": "string"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ],
                                  "title": "Result"
                                },
                                "retry_count": {
                                  "anyOf": [
                                    {
                                      "maximum": 100.0,
                                      "minimum": 0.0,
                                      "type": "integer"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ],
                                  "title": "Retry Count"
                                },
                                "screen_name": {
                                  "anyOf": [
                                    {
                                      "enum": [
                                        "welcome",
                                        "login",
                                        "register",
                                        "recovery",
                                        "verify_email",
                                        "reset_password",
                                        "account",
                                        "sessions",
                                        "materials",
                                        "chapter",
                                        "collection",
                                        "admin_policy"
                                      ],
                                      "type": "string"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ],
                                  "title": "Screen Name"
                                },
                                "status_code": {
                                  "anyOf": [
                                    {
                                      "maximum": 599.0,
                                      "minimum": 100.0,
                                      "type": "integer"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ],
                                  "title": "Status Code"
                                },
                                "target_kind": {
                                  "anyOf": [
                                    {
                                      "const": "material_content",
                                      "type": "string"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ],
                                  "title": "Target Kind"
                                },
                                "transport": {
                                  "anyOf": [
                                    {
                                      "enum": [
                                        "web",
                                        "native"
                                      ],
                                      "type": "string"
                                    },
                                    {
                                      "type": "null"
                                    }
                                  ],
                                  "title": "Transport"
                                }
                              },
                              "title": "TelemetryAttributes",
                              "type": "object"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "build": {
                          "anyOf": [
                            {
                              "pattern": "^[A-Za-z0-9._-]{1,64}$",
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "title": "Build"
                        },
                        "client_platform": {
                          "enum": [
                            "web",
                            "windows",
                            "android"
                          ],
                          "title": "Client Platform",
                          "type": "string"
                        },
                        "client_request_id": {
                          "anyOf": [
                            {
                              "format": "uuid4",
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "title": "Client Request Id"
                        },
                        "event": {
                          "maxLength": 80,
                          "pattern": "^[a-z][a-z0-9_]*(?:\\.[a-z][a-z0-9_]*){1,3}$",
                          "title": "Event",
                          "type": "string"
                        },
                        "event_id": {
                          "format": "uuid4",
                          "title": "Event Id",
                          "type": "string"
                        },
                        "level": {
                          "enum": [
                            "debug",
                            "info",
                            "warn",
                            "error",
                            "fatal"
                          ],
                          "title": "Level",
                          "type": "string"
                        },
                        "occurred_at": {
                          "format": "date-time",
                          "title": "Occurred At",
                          "type": "string"
                        },
                        "operation_id": {
                          "anyOf": [
                            {
                              "format": "uuid4",
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "title": "Operation Id"
                        },
                        "record_type": {
                          "enum": [
                            "log",
                            "analytics",
                            "performance",
                            "crash"
                          ],
                          "title": "Record Type",
                          "type": "string"
                        },
                        "release": {
                          "anyOf": [
                            {
                              "pattern": "^[A-Za-z0-9._-]{1,64}$",
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "title": "Release"
                        },
                        "request_id": {
                          "anyOf": [
                            {
                              "format": "uuid",
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "title": "Request Id"
                        },
                        "safe_stack_frames": {
                          "anyOf": [
                            {
                              "items": {
                                "type": "string"
                              },
                              "maxItems": 20,
                              "type": "array"
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "title": "Safe Stack Frames"
                        },
                        "schema_version": {
                          "const": 1,
                          "default": 1,
                          "title": "Schema Version",
                          "type": "integer"
                        }
                      },
                      "required": [
                        "event_id",
                        "record_type",
                        "event",
                        "level",
                        "occurred_at",
                        "client_platform"
                      ],
                      "title": "TelemetryEvent",
                      "type": "object"
                    },
                    "maxItems": 20,
                    "minItems": 1,
                    "title": "Events",
                    "type": "array"
                  },
                  "schema_version": {
                    "const": 1,
                    "default": 1,
                    "title": "Schema Version",
                    "type": "integer"
                  }
                },
                "required": [
                  "client_session_id",
                  "events"
                ],
                "title": "TelemetryBatchDocumented",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_TelemetryBatchRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Admin Frontend Logs",
        "tags": [
          "frontend-telemetry"
        ],
        "x-haruka-permissions": [
          "admin.login"
        ]
      }
    },
...
  }
}